# docs/android-handoff/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SwaggerExport",
    "version": "1.0"
  },
  "paths": {
...
    "/api/TienIch/PF_01_GetSysDateString": {
      "post": {
        "tags": [
          "TienIch"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PF_01_GetSysDateStringRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PF_01_GetSysDateStringRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PF_01_GetSysDateStringRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TienIch/PF_01_GetSysDate_Thang": {
      "post": {
        "tags": [
          "TienIch"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PF_01_GetSysDate_ThangRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PF_01_GetSysDate_ThangRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PF_01_GetSysDate_ThangRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TienIch/PF_02_LAY_CHI_SO_THAO_LAP": {
      "post": {
        "tags": [
          "TienIch"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PF_02_LAY_CHI_SO_THAO_LAPRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PF_02_LAY_CHI_SO_THAO_LAPRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PF_02_LAY_CHI_SO_THAO_LAPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PF_01_GetSysDateStringRequest": {
        "type": "object",
        "properties": {
          "MA_BIEN_DOC": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PF_01_GetSysDate_ThangRequest": {
        "type": "object",
        "properties": {
          "MA_BIEN_DOC": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PF_02_LAY_CHI_SO_THAO_LAPRequest": {
        "type": "object",
        "properties": {
          "MA_KHACH_HANG": {
            "type": "string",
            "nullable": true
          },
          "MA_BIEN_DOC": {
            "type": "string",
            "nullable": true
          },
          "SO_IMEI": {
            "type": "string",
            "nullable": true
          },
          "PASSWORD_K": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}